Refactor drawer state changes
This cleans up stage change logic significantly, with the intent of
making stage changes (and events) more reliable and easier to use.

Also, this allows drawers to dispatch events when they open, instead of
dispatching events when they want to open, then listening for their own
events to actually open (???). Not sure why I'd built it that way in the
first place, but it makes a bit more sense now.